Move default generator to a new module
