+++ src/sphinx_gooey/gooey.py
@@ -1,10 +1,8 @@
-import ast
-from dataclasses import dataclass
 from importlib.metadata import entry_points
 from pathlib import Path
-from textwrap import dedent
 
 from sphinx.application import Sphinx
+from sphinx.config import Config
 from sphinx.errors import ExtensionError
 from sphinx.util.logging import getLogger
 
@@ -13,38 +11,7 @@ from .directives import ExampleGallery
 logger = getLogger(__name__)
 
 
-@dataclass
-class Example:
-    path: Path
-    source_folder: Path
-    name: str = ""
-    reference: str = ""
-    summary: str = ""
-    category: str = ""
-
-    def __post_init__(self):
-        self.name = self.path.name
-        has_subdir = len(self.path.relative_to(self.source_folder.parent).parts) > 2
-        if has_subdir:
-            self.reference = f"{self.path.parts[-2]}-{self.path.stem}"
-            self.category = self.path.parts[-2]
-        else:
-            self.reference = self.path.stem
-
-        self.reference = self.reference.replace("_", "-").replace(" ", "")
-
-        mod = ast.parse(self.path.read_bytes())
-        doc = ""
-        for node in mod.body:
-            if isinstance(node, ast.Expr) and isinstance(node.value, ast.Str):
-                doc = node.value.s.strip().split("\n\n")[0].strip()
-                if not doc.endswith("."):
-                    doc += "."
-                break
-        self.summary = doc
-
-
-def generate_example_md(app: Sphinx, config):
+def generate_example_md(app: Sphinx, config: Config) -> None:
     """
     The extension configuration needs to list the source files folder and file
     extensions. This function will generate ``<filename>.md`` files for each example
@@ -69,37 +36,6 @@ def generate_example_md(app: Sphinx, config):
     automatically subdivided by subfolder in the source location.
     """
 
-    def generic(extension: str, app: Sphinx, source_folder: Path) -> list[Example]:
-        examples = []
-        for ff in source_folder.rglob(extension):
-            pth = ff.relative_to(app.srcdir)
-            if " " in str(pth):
-                logger.warning(
-                    f"The example '{pth!s}' has a space in the "
-                    "pathname which is not yet supported."
-                )
-                continue
-            example = Example(ff, source_folder)
-            examples.append(example)
-            md_file = ff.with_suffix(".md")
-            md_file.write_text(
-                dedent(
-                    f"""\
-                    ---
-                    orphan: true
-                    ---
-                    ({example.reference})=
-                    # {example.name}
-
-                    [**Source**]({ff.name})
-
-                    :::{{literalinclude}} {ff.name}
-                    :::
-                    """
-                )
-            )
-        return examples
-
     generators = app.config.sphinx_gooey_conf.pop("generators")
     for name, values in app.config.sphinx_gooey_conf.items():
         source_folder = Path(app.srcdir) / values["source"]
@@ -116,7 +52,8 @@ def generate_example_md(app: Sphinx, config):
                 jupyter = generators["jupyter"].load()
                 examples.extend(jupyter(extension, app, source_folder))
             else:
-                examples.extend(generic(extension, app, source_folder))
+                default = generators["default"].load()
+                examples.extend(default(extension, app, source_folder))
         references = set(e.reference for e in examples)
         if len(references) != len(examples):
             logger.error("There are duplicate path names in the set of examples")
@@ -124,7 +61,7 @@ def generate_example_md(app: Sphinx, config):
         app.config.sphinx_gooey_conf[name]["examples"] = examples
 
 
-def load_generator_entry_points(app: Sphinx, config):
+def load_generator_entry_points(app: Sphinx, config: Config):
     app.config.sphinx_gooey_conf["generators"] = entry_points(
         group="sphinx_gooey.generators"
     )
@@ -136,6 +73,9 @@ def setup(app: Sphinx):
     app.add_config_value("sphinx_gooey_conf", {}, "html")
     app.add_directive("example-gallery", ExampleGallery)
 
+    # If myst_nb is set up after myst_parser there are conflicts about common
+    # configuration options and directives. So we can't set up myst_nb only when
+    # the Jupyter stuff is loaded, it has to come here.
     try:
         app.setup_extension("myst_nb")
     except ExtensionError:
